Move download utility functions from Utils to separate class
